+++ src/cake/tasks/clean.clj
@@ -2,10 +2,14 @@
   (:use cake cake.core cake.ant)
   (:import (org.apache.tools.ant.taskdefs Delete Mkdir)))
 
+(defn clean-dir [dir]
+  (ant Delete {:include-empty-dirs true}
+       (add-fileset {:dir dir :includes "**/*"})))
+
 (deftask clean
   "Remove cake build artifacts."
-  (let [files ["pom.xml" "classes" "lib" "build"]]
-    (doseq [file (map file (concat files (:clean *project*)))]
-      (if (.isDirectory file)
-        (ant Delete {:dir  file})
-        (ant Delete {:file file})))))
+  (doseq [dir ["classes" "build"]]
+    (clean-dir dir))
+  (when (= ["deps"] (:clean *opts*))
+    (clean-dir "lib")
+    (ant Delete {:file "pom.xml"})))