+++ test/test_core.clj
@@ -2,19 +2,20 @@
   (:use clojure.test cake cake.core))
 
 (defmacro with-project [bindings & body]
-  `(binding [*project* {:name "project-stub", :version "0.0.0", :root "/home/project"}]
+  `(binding [*project* {:name "project-stub", :version "0.0.0"}
+             *root*    "/home/project"]
      (let ~bindings
        ~@body)))
 
 (deftest file-fn
   (testing "single string"
     (with-project [s "foo/bar/baz"]
-      (is (= (str (:root *project*) "/" s)
+      (is (= (str *root* "/" s)
              (.toString (file s))))))
   
   (testing "multiple strings"
     (with-project [a "foo", b "bar"]
-      (is (= (str (:root *project*) "/" a "/" b)
+      (is (= (str *root* "/" a "/" b)
              (.toString (file a b))))))
 
   (testing "single file"
@@ -23,7 +24,7 @@
   
   (testing "file and string"
     (with-project [foo "foo", f (file foo), s "bar"]
-      (is (= (str (:root *project*) "/" foo "/" s)
+      (is (= (str *root* "/" foo "/" s)
              (.toString (file f s))))))
 
   (testing "tilde expansion"
@@ -33,5 +34,5 @@
 
   (testing "no arguments"
     (with-project [p "/foo/bar"]
-      (is (= (str (:root *project*))
+      (is (= (str *root*)
              (.toString (file)))))))