+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### `0.4.3` - August 20, 2010
+- switch to clojure 1.2.0
+- support restarting of project jvm from clojure using `cake.core/bake-restart` function
+- restart project jvm after fetching new deps or compiling new `.class` files
+- make `deps` a dependency of `compile`
+- fix bug that was leaving defunct jvms running
+- only fetch deps if `project.clj` is newer than a file in `lib` or `pom.xml`
+- add force option to compile and deps
+- change `clean` to not remove deps; use `clean deps` to remove deps too
+- allow `invoke` function to override `*opts*`, though a given task will still only be executed once
+- improved debugging messages for java compilation
+
 ### `0.4.0` - August 18, 2010
 - full Windows support, including repl and readline; [install instructions](http://wiki.github.com/ninjudd/cake/cake-on-windows)
 - moved environment-specific vars to `cake` namespace and make them available to all commands (e.g. `cake/*env*`, `cake/*pwd*`)